requirements.txt
# WNAP - Web Novel Archive Pipeline
# Python 3.10+ required

# GUI
customtkinter>=5.2.0

# Archive handling
py7zr>=0.20.0
rarfile>=4.0

# Web scraping (for GenreClassifier)
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0

# Security
cryptography>=40.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
hypothesis>=6.0.0

# Packaging (optional)
pyinstaller>=6.0.0
